cleaned up run_first_proc() code
moved run_first_proc() into kernel/exec.c and wrote iret_to_userspace
in arch/x86/cpu.c. This should probably be moved out of cpu.c and
into a different file.

+++ kernel/kernel/proc.c
@@ -35,8 +35,8 @@ void proc_init(struct proc_struct *proc) {
 void thread_init(struct thread_struct *thread) {
   memset(thread->kstack, 0, THREAD_KSTACK_SIZE);
 
-  thread->kstack_hi = thread->kstack + THREAD_KSTACK_SIZE;
-  thread->kstack_lo = thread->kstack;
+  thread->kstack_hi = (size_t) (thread->kstack + THREAD_KSTACK_SIZE);
+  thread->kstack_lo = (size_t) (thread->kstack);
 
   list_elem_init(thread, thread_link);
 }
